[FIX] stock_barcodes: Move qty_available field from read base to read picking TT51889
