+++ stock_barcodes/wizard/stock_barcodes_read_picking.py
@@ -6,7 +6,7 @@ from collections import OrderedDict, defaultdict
 from odoo import _, api, fields, models
 from odoo.exceptions import ValidationError
 from odoo.fields import first
-from odoo.tools.float_utils import float_compare
+from odoo.tools.float_utils import float_compare, float_round
 from odoo.tools.safe_eval import safe_eval
 
 _logger = logging.getLogger(__name__)
@@ -78,6 +78,7 @@ class WizStockBarcodesReadPicking(models.TransientModel):
     company_id = fields.Many2one(related="picking_id.company_id")
     todo_line_is_extra_line = fields.Boolean(related="todo_line_id.is_extra_line")
     forced_todo_key = fields.Char()
+    qty_available = fields.Float(compute="_compute_qty_available")
 
     @api.depends("todo_line_id")
     def _compute_todo_line_display_ids(self):
@@ -114,6 +115,40 @@ class WizStockBarcodesReadPicking(models.TransientModel):
             for line in product_moves:
                 rec.total_product_uom_qty += line.product_uom_qty
                 rec.total_product_qty_done += line.quantity_done
+
+    @api.depends("location_id", "product_id", "lot_id")
+    def _compute_qty_available(self):
+        if not self.product_id or self.location_id.usage != "internal":
+            self.qty_available = 0.0
+            return
+        domain_quant = [
+            ("product_id", "=", self.product_id.id),
+            ("location_id", "=", self.location_id.id),
+        ]
+        if self.lot_id:
+            domain_quant.append(("lot_id", "=", self.lot_id.id))
+        # if self.package_id:
+        #     domain_quant.append(('package_id', '=', self.package_id.id))
+        groups = self.env["stock.quant"].read_group(
+            domain_quant, ["quantity"], [], orderby="id"
+        )
+        self.qty_available = groups[0]["quantity"]
+        # Unexpected done quantities must reduce qty_available
+        if self.lot_id:
+            done_move_lines = self.move_line_ids.filtered(
+                lambda m: m.product_id == self.product_id and m.lot_id == self.lot_id
+            )
+        else:
+            done_move_lines = self.move_line_ids.filtered(
+                lambda m: m.product_id == self.product_id
+            )
+        for sml in done_move_lines:
+            over_done_qty = float_round(
+                sml.qty_done - sml.product_uom_qty,
+                precision_rounding=sml.product_uom_id.rounding,
+            )
+            if over_done_qty > 0.0:
+                self.qty_available -= over_done_qty
 
     def name_get(self):
         return [
